+++ Realm/RLMSyncConfiguration_Private.hpp
@@ -36,12 +36,11 @@ NS_ASSUME_NONNULL_BEGIN
                     realmURL:(NSURL *)url
                customFileURL:(nullable NSURL *)customFileURL
                    isPartial:(BOOL)isPartial
-                  stopPolicy:(RLMSyncStopPolicy)stopPolicy
-                errorHandler:(std::function<realm::SyncSessionErrorHandler>)errorHandler;
+                  stopPolicy:(RLMSyncStopPolicy)stopPolicy;
 
 - (instancetype)initWithRawConfig:(realm::SyncConfig)config;
 
-- (realm::SyncConfig)rawConfiguration;
+- (realm::SyncConfig&)rawConfiguration;
 
 @end